Feature(integration): Facebook Pixel Mandatory properties fix (#1307)
* Feature(integration): Facebook Pixel Mandatory properties fix

* Update v0/destinations/facebook_pixel/transform.js


* Feature(integration): Facebook Pixel  edits

* Feature(integration): Facebook Pixel adding fbc support

* Feature(integration): Facebook Pixel editing error condition for prod list viewed

* Update v0/destinations/facebook_pixel/transform.js


* Feature( integration): Facebook Pixel review comments addressed

* Feature( integration): Facebook Pixel review comments addressed

* Feature( integration): Facebook Pixel review comments addressed

* Feature( integration): Facebook Pixel review comments addressed

* Feature( integration): Facebook Pixel review comments addressed
